Move sqlite to keystone internals
