+++ core/schema/sqlite.capnp
