--- modules/sqlite-ks/Cargo.toml
@@ -1,35 +1,0 @@
-[package]
-name = "sqlite-ks"
-version = "0.1.0"
-edition.workspace = true
-rust-version.workspace = true
-license.workspace = true
-authors = ["Erik McClure <erikm@fundament.software>"]
-description = """
-Keystone module that embeds sqlite and exposes sqlite database capabilities.
-"""
-homepage = "https://github.com/Fundament-Software/keystone/tree/main/modules/sqlite-ks"
-repository = "https://github.com/Fundament-Software/keystone/tree/main/modules/sqlite-ks"
-readme = "README.md"
-keywords = ["backend", "infrastructure", "sqlite", "distributed"]
-
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[dependencies]
-color-eyre.workspace = true
-eyre.workspace = true
-cap-std.workspace = true
-tokio.workspace = true
-capstone.workspace = true
-capstone-rpc.workspace = true
-capstone-macros.workspace = true
-capstone-import.workspace = true
-async-backtrace.workspace = true
-#sqlite_nostd = { git="https://github.com/vlcn-io/sqlite-rs-embedded.git" }
-rusqlite.workspace = true
-rand.workspace = true
-
-#[features]
-#default = ["static"]
-#static = ["sqlite_nostd/static"]
