--- modules/sqlite-ks/rustfmt.toml
@@ -1,1 +1,0 @@
-max_width = 200
